Change amount off attribute to integer instead of decimal

+++ db/migrate/20190601233240_create_coupons.rb
@@ -3,7 +3,7 @@ class CreateCoupons < ActiveRecord::Migration[5.1]
     create_table :coupons do |t|
       t.string :name
       t.string :code
-      t.decimal :amount_off
+      t.integer :amount_off
       t.boolean :active, default: true
       t.references :user, foreign_key: true
     end
